+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - batch processing for OpenAI
 - graph analytics CLI commands, including an analytics report (what is the most important chunk in this document...?)
 - graph conversion CLI command to take the `graph.json` to JSONLD, GML, GraphML and more.
+- AWS Neptune import script and documentation
 
 ## v0.4.1